requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
python-multipart==0.0.22
mongoengine==0.29.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0